fix copilot header imports
