+++ src/agents/copilot-dynamic-headers.ts
@@ -1,7 +1,30 @@
 import type { Context } from "@earendil-works/pi-ai";
-import { COPILOT_INTEGRATION_ID, buildCopilotIdeHeaders } from "../plugin-sdk/provider-auth.js";
 
-export { COPILOT_INTEGRATION_ID, buildCopilotIdeHeaders } from "../plugin-sdk/provider-auth.js";
+/** @deprecated GitHub Copilot provider-owned helper; do not use from third-party plugins. */
+export const COPILOT_EDITOR_VERSION = "vscode/1.107.0";
+/** @deprecated GitHub Copilot provider-owned helper; do not use from third-party plugins. */
+export const COPILOT_USER_AGENT = "GitHubCopilotChat/0.35.0";
+/** @deprecated GitHub Copilot provider-owned helper; do not use from third-party plugins. */
+export const COPILOT_EDITOR_PLUGIN_VERSION = "copilot-chat/0.35.0";
+/** @deprecated GitHub Copilot provider-owned helper; do not use from third-party plugins. */
+export const COPILOT_GITHUB_API_VERSION = "2025-04-01";
+/** @deprecated GitHub Copilot provider-owned helper; do not use from third-party plugins. */
+export const COPILOT_INTEGRATION_ID = "vscode-chat";
+
+/** @deprecated GitHub Copilot provider-owned helper; do not use from third-party plugins. */
+export function buildCopilotIdeHeaders(
+  params: {
+    includeApiVersion?: boolean;
+  } = {},
+): Record<string, string> {
+  return {
+    "Accept-Encoding": "identity",
+    "Editor-Version": COPILOT_EDITOR_VERSION,
+    "Editor-Plugin-Version": COPILOT_EDITOR_PLUGIN_VERSION,
+    "User-Agent": COPILOT_USER_AGENT,
+    ...(params.includeApiVersion ? { "X-Github-Api-Version": COPILOT_GITHUB_API_VERSION } : {}),
+  };
+}
 
 function inferCopilotInitiator(messages: Context["messages"]): "agent" | "user" {
   const last = messages[messages.length - 1];
